Rename everything for uniqueness
I think OpenAM is using the DuoNode classname (registered in the plugin's getNodesByVersion method) as a unique identifier, so you can't have both show up in the UI since the v2 SDK plugin is also DuoNode.

Here's the JSON the auth tree UI gets to populate the node sidebar:

```
      {
         "_id":"DuoNode",
         "name":"Duo Node",
         "collection":true
      }
```

So I'm hoping renaming it sorts that out!